Create parameter table

Creates a table with the final estimates and percentage CV for all parameters in an nlmixr output file. This can be done for one or multiple models for easy comparison

Usage

par_table(
  proj,
  models,
  outnm = NULL,
  projloc = ".",
  bsv = FALSE,
  shrink = FALSE,
  backt = FALSE,
  formatting = FALSE,
  ...
)

Arguments

proj

project object

models

character vector with model names to create table for

outnm

character with name of the output file (see details)

projloc

character with the base location of the shinyMixR project

bsv

logical indicating if between subject variability (BSV) should be added to table

shrink

logical indicating if shrinkage should be added to table

backt

logical indicating if the backtransformed parameters should be returned oposed to the original values

formatting

logical indicating if the formatting should be applied to present the table (not implemented for latex output)

...

additional arguments passed to ltx_plot or html_plot

Value

in case no outnm is defined a data frame will be returned otherwise the results are saved to disk

Details

In case a model is saved, a directory with the name of the model is created within the analysis folder of the current project. Then within this folder the file is saved as outnm. This method was chosen so the interface can easily index applicable files for a certain model. However, this means that output is alwasy saved in this directly regardless of the location of outnm In case multiple models are selected the result will be written to the name of the first model in the models vector.

Author

Richard Hooijmaijers

Examples

if (FALSE) { # \dontrun{
 par_table(proj,"run1")
} # }