Runs a nlmixr model

Runs an nlmixr model from a project object with the possibility to run in an external rsession using a system call (tested within linux only)

Usage

run_nmx(
  mod,
  proj = proj,
  ext = TRUE,
  saverds = TRUE,
  autoupdate = TRUE,
  projloc = ".",
  addcwres = TRUE,
  addnpde = TRUE
)

Arguments

mod

character with the model file present in project object

proj

project object

ext

logical indicating if the model should be run external in a separate r session

saverds

logical indicating if the model results should be saved in a rds file

autoupdate

logical indicating if the project object should automatically update

projloc

character with the base location of the shinyMixR project

addcwres

logical indicating if CWRES should be added to the output

addnpde

logical indicating if NPDE should be added to the output

Value

In case the model is not submitted in a separate R session, the results from nlmixr are returned otherwise the result of the system call will be returned

Details

the meta data is obtained by compiling the model. The dataset, estimation method and control list are then included in the nlmixr call. Meta data is included in the model function which is comparable with NONMEM. This method was chosen so that all information to run a model is kept togehter in one function

Author

Richard Hooijmaijers

Examples

if (FALSE) { # \dontrun{
 run_nmx("run1",proj)
} # }